refactor: [#69] Bookmark ShareInfo Controller 분리
