+++ docker-compose.prod.yml
@@ -5,7 +5,7 @@ services:
     container_name: mysql
     image: mysql:latest
     ports:
-      - "3306:3306"
+      - "3307:3307"
     restart: always
     environment:
       MYSQL_DATABASE: ${DB_DATABASE}
